Enable the user to bind the client to realm roles directly
Using the 'roles' key would create a new client role that would be wrapped in the listed realm roles as composites. This enable
us to define a new role and grant it so, let's say, all admins in the system. But this does not enable us to define that a given
client has to receive a given realm role in order to operate. This new feature now enables the later.

This is important so we can have automated/reproducible keycloak realm configuration