ClientScopeController now lets its UUID be generated by Keycloak.
It now sets the client scope _name_ rather than _id_.

Not setting the name was causing issues on Keycloak 7.0.0.
Also, the Kubernetes resource field used was called `name`,
so it seems like a mistake that it was really setting the UUID.
Not to mention setting the UUID manually seems like bad practice in
general.